refactor integration tests package
