+++ integration/s3_integration_test.go
@@ -1,4 +1,4 @@
-package integration_test
+package integration
 
 import (
 	"context"
