+++ integration/udf.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package testutils
+package integration
 
 const UDF = `
 local function putBin(r,name,value)
